upstream-layers/bitbake/lib/toaster/tests/toaster-tests-requirements.txt
selenium>=4.13.0
pytest==7.4.2
pytest-django==4.5.2
pytest-env==1.1.0
pytest-html==4.0.2
pytest-metadata==3.0.0
pytest-order==1.1.0
requests

